requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
astor=0.8.1=pypi_0
atari-py=0.2.6=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
bleach=3.1.0=pypi_0
box2d-py=2.3.8=pypi_0
ca-certificates=2020.1.1=0
cachetools=4.0.0=pypi_0
certifi=2019.11.28=py37_0
cffi=1.13.2=pypi_0
chardet=3.0.4=pypi_0
click=7.0=pypi_0
cloudpickle=1.2.2=pypi_0
configparser=4.0.2=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.14=pypi_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=pypi_0
defusedxml=0.6.0=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=pypi_0
expat=2.2.6=he6710b0_0
fasteners=0.15=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
gast=0.2.2=pypi_0
gitdb2=2.0.6=pypi_0
gitpython=3.0.5=pypi_0
glfw=1.9.1=pypi_0
glib=2.63.1=h5a9c865_0
google-auth=1.10.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.8=pypi_0
gql=0.2.0=pypi_0
graphql-core=1.1=pypi_0
grpcio=1.26.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gym=0.15.4=dev_0
h5py=2.10.0=pypi_0
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
imageio=2.6.1=pypi_0
importlib-metadata=1.3.0=pypi_0
intel-openmp=2019.4=243
ipykernel=5.1.3=pypi_0
ipython=7.11.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
jedi=0.15.2=pypi_0
jinja2=2.10.3=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=5.3.4=pypi_0
jupyter-console=6.0.0=pypi_0
jupyter-core=4.6.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.1.0=py37he6710b0_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
lockfile=0.12.2=pypi_0
markdown=3.1.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.2=pypi_0
mistune=0.8.4=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mlagents=0.13.1=pypi_0
mlagents-envs=0.13.1=pypi_0
monotonic=1.5=pypi_0
more-itertools=8.0.2=pypi_0
mujoco-py=1.50.1.68=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.3=pypi_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
notebook=6.0.2=pypi_0
numpy=1.18.1=pypi_0
numpy-base=1.17.4=py37hde5b4d6_0
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
opencv-python=4.1.2.30=pypi_0
openssl=1.1.1d=h7b6447c_3
opt-einsum=3.1.0=pypi_0
pandas=0.25.3=py37he6710b0_0
pandocfilters=1.4.2=pypi_0
parso=0.5.2=pypi_0
patchelf=0.10=he6710b0_0
pathtools=0.1.2=pypi_0
patsy=0.5.1=py37_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.0.0=py37hb39fc2d_0
pip=19.3.1=py37_0
prometheus-client=0.7.1=pypi_0
promise=2.3=pypi_0
prompt-toolkit=2.0.10=pypi_0
protobuf=3.11.2=pypi_0
psutil=5.6.7=pypi_0
ptyprocess=0.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycparser=2.19=pypi_0
pyglet=1.3.2=pypi_0
pygments=2.5.2=pypi_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.15.7=pypi_0
python=3.7.4=h265db76_1
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
pytz=2019.3=py_0
pyyaml=5.3=pypi_0
pyzmq=18.1.1=pypi_0
qt=5.9.7=h5867ecd_1
qtconsole=4.6.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scipy=1.4.1=pypi_0
seaborn=0.9.0=pyh91ea838_1
send2trash=1.5.0=pypi_0
sentry-sdk=0.14.1=pypi_0
setuptools=44.0.0=py37_0
shortuuid=0.5.0=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
smmap2=2.0.5=pypi_0
sqlite=3.30.1=h7b6447c_0
statsmodels=0.10.1=py37hdd07704_0
subprocess32=3.5.4=pypi_0
tensorboard=2.0.2=pypi_0
tensorflow=2.0.0=pypi_0
tensorflow-estimator=2.0.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tk=8.6.8=hbc83047_0
torchvision=0.5.0=py37_cu101
tornado=6.0.3=py37h7b6447c_0
tqdm=4.42.0=py_0
traitlets=4.3.3=pypi_0
urllib3=1.25.7=pypi_0
wandb=0.8.24=pypi_0
watchdog=0.10.1=pypi_0
wcwidth=0.1.8=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.0=pypi_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.11.2=pypi_0
xz=5.2.4=h14c3975_4
zipp=0.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
